restructure the add menu to be a pop-out overlay
This is going to be the start.  More to be done but committing the code
now and going home.

Things still to be done are the replace player and endpoint to add a
player and remove the other one to FAs.  Release Player slide out menu.

Start to look into trade proposals.